+++ crates/hir-def/src/expr_store/pretty.rs
@@ -771,6 +771,8 @@ impl Printer<'_> {
                     ast::UnaryOp::Deref => "*",
                     ast::UnaryOp::Not => "!",
                     ast::UnaryOp::Neg => "-",
+                    ast::UnaryOp::BigAnd => "&&&",
+                    ast::UnaryOp::BigOr => "|||",
                 };
                 w!(self, "{}", op);
                 self.print_expr_in(prec, *expr);